Fix type conversion for retry multiplier
Convert float32 to float64 when assigning retry multiplier from
config to avoid type mismatch error.

+++ pkg/ai/client.go
@@ -157,7 +157,7 @@ func convertAIConfigToServiceConfig(cfg config.AIConfig) *AIServiceConfig {
 		retryConfig.MaxDelay = cfg.Retry.MaxDelay.Duration
 	}
 	if cfg.Retry.Multiplier > 0 {
-		retryConfig.BackoffMultiplier = cfg.Retry.Multiplier
+		retryConfig.BackoffMultiplier = float64(cfg.Retry.Multiplier)
 	}
 	retryConfig.Jitter = cfg.Retry.Jitter
